requirements.txt
pillow~=10.2.0
pyyaml~=6.0.1
zlib~=1.2.13
gymnasium~=0.29.1
numpy~=1.26.4
stable-baselines3~=2.3.2
imageio~=2.34.1
colorist~=1.8.1
matplotlib~=3.9.2
setuptools~=69.1.1
readline~=8.2
torchvision~=0.15.1
tensorboard~=2.11.2